test: sync_main_flow 并入 test_sync.py
